+++ server.js
@@ -1,86 +1,30 @@
 import express from "express"
 import bodyParser from "body-parser"
-import { getNodeStatus, shutDownNode, startNode } from "./src/nodeCommand.js"
-
+import { getNodeStatus, shutDownNode, startNode } from "./src/rpc.js"
+import { createWebsocketChannel } from './src/websocket.js'
 import { createServer } from "http";
-import { Server } from "socket.io";
+
 
 const app = express()
 const clientApp = express()
 const port = 5000
-
 let httpServer = createServer(app) 
-const io = new Server(httpServer, {})
-
-
-import fs from "fs"
-import request from "request"
-import {checkStacksNodeExists} from './utils/stacksNode.js'
-import {selectSystem} from './utils/utils.js'
-
-io.on('connection', (socket) => {
-    socket.on("download", async (msg ,b) => {
-        console.log(msg, b)
-        let exists = await checkStacksNodeExists()
-        console.log(exists)
-        if (exists) {
-            return;
-        }
-        let system = selectSystem()
-        let file_url = 'https://github.com/Daemon-Technologies/Mining-Bot/releases/download/1.0.0/stacks-node-'
-        switch (system){
-            case "darwin": file_url+="mac"
-                           break;
-            case "linux":  file_url+=system
-                           break;
-            case "default": return "Do not support other system except Darwin and linux";
-        }
-        let out = fs.createWriteStream('stacks-node')
-        
-        let req = request({
-            method: 'GET',
-            uri: file_url
-        })
-        let cur = 0;
-        let total = 0;
-        let percent = 0;
-        let lastPercent = 0;
-        req.pipe(out);
-
-        req.on('data', function (chunk) {
-            cur += chunk.length
-            if (total != 0){
-                percent = cur/total
-            }
-            if (percent - lastPercent > 0.01){
-                lastPercent = percent
-                io.emit("download_info", lastPercent)
-            }    
-        });
-
-        req.on('end', function() {
-            //console.log("say something")
-            io.emit("download_info", lastPercent)
-            io.emit("download_complete", 1)
-            percent = 1
-        });
-        
-        req.on( 'response', function (data) {
-            total = data.headers[ 'content-length' ];
-            console.log("response:", data.headers[ 'content-length' ] );
-        });
-
-        req.on('error', function (err){
-            console.log(err)
-        })
-    });
-    console.log("user connected");
-})
-
 
 app.use(bodyParser.json());
 
 app.all("*", function (req, res, next) {
+    res.header("Access-Control-Allow-Origin", req.headers.origin || '*');
+    res.header("Access-Control-Allow-Headers", "Content-Type, Authorization, X-Requested-With");
+    res.header("Access-Control-Allow-Methods", "PUT,POST,GET,DELETE,OPTIONS");
+    res.header("Access-Control-Allow-Credentials", true);
+    if (req.method == 'OPTIONS') {
+      res.sendStatus(200);
+    } else {
+      next();
+    }
+})
+
+clientApp.all("*", function (req, res, next) {
     res.header("Access-Control-Allow-Origin", req.headers.origin || '*');
     res.header("Access-Control-Allow-Headers", "Content-Type, Authorization, X-Requested-With");
     res.header("Access-Control-Allow-Methods", "PUT,POST,GET,DELETE,OPTIONS");
@@ -134,3 +78,5 @@ clientApp.use(express.static('dist'));
 clientApp.get('/*', function (req, res) {
   res.sendFile('dist/index.html', { root: '.' });
 });
+
+createWebsocketChannel(httpServer)
