+++ src/rpc.js
@@ -1,71 +1,10 @@
 import child_process from "child_process"
 import fs from "fs"
 import execa from "execa"
-import crypto from "crypto"
-import os from "os"
 import request from "request"
-
-function splitProcess(commands){
-    //PID TTY TIME CMD
-    const Verbose = false;
-    let result = []
-    if (Verbose) console.log(typeof(commands))
-
-    let commandArray = commands.split('\n')
-    if (Verbose) console.log(commandArray)
-    for (let i in commandArray){
-        let item = commandArray[i]
-        if (item.length < 10) continue;
-        // split header space
-        let index = 0
-        let command = {'PID': '', 'TTY':'' ,'TIME':'' ,'CMD':''}
-        while (index < 4){
-            if (Verbose) console.log(item, item.search(" "))
-            if (item.search(" ") == 0){
-                item = item.slice(1)
-                if (Verbose) console.log(item)
-                continue;
-            }
-            else{
-                let nextPart = item.search(" ")
-                let t = item.slice(0, nextPart)
-                if (Verbose) console.log("Debug:",item, nextPart, t)                
-                switch (index){
-                    case 0 : command.PID = t; break;
-                    case 1 : command.TTY = t; break;
-                    case 2 : command.TIME = t; break;
-                    case 3 : command.CMD = item; break;
-                }
-                item = item.slice(nextPart)
-                if (Verbose) console.log("Debug After:",item, nextPart, t)
-                index++;
-                continue;
-            }
-        }
-        result.push(command)
-        if (Verbose) console.log(command)
-        
-    }
-    return result
-}
-
-function replaceSegment(keyword, value ,strFile){
-    const Verbose = false;
-    let start = strFile.search(keyword)
-    let end = start
-    while (end <= strFile.length){
-        if (strFile.charAt(end) == '\n')
-            break;
-        end++
-    }
-    if (Verbose) console.log(end, strFile[end])
-    
-    let injection = `${keyword} = ${value}\n`;
-    if (Verbose) console.log("injection", injection)
-    let result = strFile.slice(0, start) + injection + strFile.slice(end - strFile.length + 1) 
-
-    return result
-}
+import { selectSystem} from '../utils/utils.js'
+import { checkStacksNodeMD5, getStacksNodeMD5, checkStacksNodeExists, deleteStacksNode, replaceSegment, getMinerAddress, isNodeStart } from '../utils/stacksNode.js'
+import { splitProcess } from '../utils/sysCommand.js'
 
 function updateMinerToml(data){
     const {seed, burn_fee_cap, network} = data
@@ -103,94 +42,6 @@ function updateMinerToml(data){
     
 }
 
-async function getMinerAddress(){
-    let strFile = fs.readFileSync("./Miner.txt", 'utf-8');
-    return strFile;
-}
-
-async function isNodeStart(commands){
-    const Verbose = false;
-    const key_word = "stacks-node start"
-    
-    let address = await getMinerAddress()
-
-    for (let index in commands){
-        if (Verbose) console.log(commands[index].CMD.search(key_word))
-        if (commands[index].CMD.search(key_word) != -1)
-            return {status: true, PID: commands[index].PID, address: address}
-    }
-    return {status: 500, PID: -5}
-}
-
-
-
-function selectSystem(){
-    return os.platform()
-}
-
-function checkStacksNodeExists(){
-    return new Promise(function(resolve){
-        try{
-            let stat = fs.statSync('./stacks-node').isFile()
-           // console.log("success:", stat)
-            resolve(true)
-        }
-        catch(error){  
-            //console.log("no stacks-node found error:",error)
-            resolve(false)
-        }
-    })
-    
-}
-
-function getStacksNodeMD5(){
-    
-    return new Promise(function(resolve){
-        if (!checkStacksNodeExists()) resolve(false);
-        let rs = fs.createReadStream('./stacks-node');
-        
-        let hash = crypto.createHash('md5');
-        rs.on('data', hash.update.bind(hash));
-            rs.on('end', function () {
-                let result = hash.digest('hex')
-                //console.log("in here:", result);
-                resolve(result);
-            });   
-        });
-}
-
-async function checkStacksNodeMD5(){
-    let md5Result = await getStacksNodeMD5()
-    //console.log("md5Result:",md5Result)
-    let system = selectSystem()
-    let darwin_md5 = "c93d51447dfd5a1c737dc1717117b933"
-    let linux_md5 = "5557c6ebdfded15e3caef6e3dc7d4643"
-    switch (system){
-        case 'darwin': if (md5Result != darwin_md5) 
-                            return false;
-                       break;
-        case 'linux': if (md5Result != linux_md5) 
-                            return false;
-                      break;
-        default : return false;
-    }
-    return true;
-}
-
-function deleteStacksNode(){
-    return new Promise(function(resolve){
-        try {
-            fs.unlinkSync('./stacks-node')
-            //file removed
-            resolve(true)
-            //return { status : 401, data : "Stacks-node doesn't complete, please re-download it"}
-        } catch(err) {
-            console.error(err)
-            resolve(false)
-            //return { status : 402, data : "Stacks-node doesn't complete, but there is some issue when deleting it. Please delete manually, it is in the folder of Mining-Local-Server."}
-        } 
-    })
-}
 
 /* nodeStatus 
     -1 no Mining-Local-Server Started
@@ -271,6 +122,12 @@ function importaddressRPC(address){
 }
 
 export async function startNode(data){
+
+    //2021.1.3 support debug mode & 
+    // const { seed, burn_fee_cap, network, address, debug } = data
+
+    // Check Encrypted Message
+    
     const Verbose = true
 
     // Check stacks-node exists
