+++ utils/sysCommand.js
@@ -1,0 +1,45 @@
+export function splitProcess(commands){
+    //PID TTY TIME CMD
+    const Verbose = false;
+    let result = []
+    if (Verbose) console.log(typeof(commands))
+
+    let commandArray = commands.split('\n')
+    if (Verbose) console.log(commandArray)
+    for (let i in commandArray){
+        let item = commandArray[i]
+        if (item.length < 10) continue;
+        // split header space
+        let index = 0
+        let command = {'PID': '', 'TTY':'' ,'TIME':'' ,'CMD':''}
+        while (index < 4){
+            if (Verbose) console.log(item, item.search(" "))
+            if (item.search(" ") == 0){
+                item = item.slice(1)
+                if (Verbose) console.log(item)
+                continue;
+            }
+            else{
+                let nextPart = item.search(" ")
+                let t = item.slice(0, nextPart)
+                if (Verbose) console.log("Debug:",item, nextPart, t)                
+                switch (index){
+                    case 0 : command.PID = t; break;
+                    case 1 : command.TTY = t; break;
+                    case 2 : command.TIME = t; break;
+                    case 3 : command.CMD = item; break;
+                }
+                item = item.slice(nextPart)
+                if (Verbose) console.log("Debug After:",item, nextPart, t)
+                index++;
+                continue;
+            }
+        }
+        result.push(command)
+        if (Verbose) console.log(command)
+        
+    }
+    return result
+}
+
+
